refactor!: change how the user interacts with the API.
This commit will change files that interact with the Wallhaven API.
All functions that were in `api.py` are now inside a class called Api.
A new exception called `APIException` was created. This exception is
only raised when the API returns a status code of 401, 404 or 429.

BREAKING CHANGES: Now users have to import and instantiate
`wallhaven.Api` to interact with the Wallhaven API.